First pass to separate folders
