+++ .gitignore
@@ -6,6 +6,7 @@
 /.pnp
 .pnp.js
 /venv
+node_modules/
 
 # testing
 /coverage
