+++ backend/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "spektate",
+  "version": "0.1.0",
+  "main": "build/dist/index.js",
+  "types": "build/dist/index.d.ts",
+  "license": "MIT",
+  "dependencies": {
+    "@types/express": "^4.17.0",
+    "spektate": "^1.0.10",
+    "body-parser": "^1.19.0"
+  },
+  "scripts": {
+    "start": "ts-node src/index.ts",
+    "build": "tsc src/** --outDir ../build/backend/",
+    "lint": "tslint -p tsconfig.json -c tslint.json src/*.ts"
+  },
+  "devDependencies": {
+    "@types/jest": "^24.0.18",
+    "@types/jest-when": "^2.7.0",
+    "@types/node": "^12.6.8",
+    "lint-staged": "^9.2.5",
+    "prettier": "^1.18.2",
+    "ts-jest": "^24.0.2",
+    "ts-node": "^8.5.4",
+    "tslint": "^5.20.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "^3.5.3"
+  }
+}
