+++ backend/src/server.ts
@@ -1,9 +1,9 @@
 import * as express from "express";
 import * as path from "path";
-import { get as authorGet } from "./backend/author";
-import { get as clusterSyncGet } from "./backend/clustersync";
-import { get as deploymentGet } from "./backend/deployment";
-import { get as prGet } from "./backend/pullrequest";
+import { get as authorGet } from "./author";
+import { get as clusterSyncGet } from "./clustersync";
+import { get as deploymentGet } from "./deployment";
+import { get as prGet } from "./pullrequest";
 
 const app = express();
