+++ docker-entrypoint.sh
@@ -8,4 +8,5 @@ do
     export $key=$value
 done
 
+cd backend
 node server.js
