+++ frontend/package.json
@@ -4,23 +4,19 @@
   "main": "build/dist/index.js",
   "types": "build/dist/index.d.ts",
   "proxy": "http://localhost:8001",
+  "license": "MIT",
   "dependencies": {
-    "@types/express": "^4.17.0",
     "azure-devops-ui": "^2.166.1",
-    "cli-table": "^0.3.1",
-    "commander": "^2.20.0",
     "office-ui-fabric-react": "^7.10.0",
     "open": "^6.4.0",
     "react": "^16.8.5",
     "react-dom": "^16.8.5",
     "react-scripts-ts": "^4.0.8",
-    "spektate": "^1.0.10",
-    "body-parser": "^1.19.0"
+    "spektate": "^1.0.10"
   },
   "scripts": {
-    "start-backend": "ts-node src/backend/index.ts",
     "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build && tsc src/backend/** --outDir build/backend/ && tsc src/server.ts --outDir build/",
+    "build": "react-scripts-ts build && mv build ../build",
     "eject": "react-scripts-ts eject",
     "lint": "tslint -p tsconfig.json -c tslint.json src/**/*.ts",
     "refresh": "rm -rf ./node_modules ./package-lock.json && npm install"
@@ -31,7 +27,6 @@
     "@types/node": "^12.6.8",
     "@types/react": "^16.8.10",
     "@types/react-dom": "^16.8.3",
-    "chai": "^4.2.0",
     "husky": "^3.0.4",
     "lint-staged": "^9.2.5",
     "pkg": "^4.4.0",
