+++ frontend/src/assets/pipeline-proto.png
